refactor: split logic unrelated to authentication domain
- 유저 도메인에 가까운 프로필 등록로직을, 인증인가 도메인에서 분리한다
- 유저 도메인에 가까운 닉네임 중복체크 로직을, 인증인가 도메인에서 분리한다
