+++ src/main/java/com/example/univeus/domain/auth/service/AuthService.java
@@ -1,7 +1,5 @@
 package com.example.univeus.domain.auth.service;
 
-import com.example.univeus.presentation.auth.dto.request.AuthRequest.Nickname;
-import com.example.univeus.presentation.auth.dto.request.AuthRequest.Profile;
 import com.example.univeus.presentation.auth.dto.response.AuthResponse.ResponseTokens;
 
 public interface AuthService {
@@ -10,8 +8,4 @@ public interface AuthService {
     ResponseTokens issueTokens(Long memberId);
 
     ResponseTokens createOrLogin(String uri);
-
-    void registerProfile(Long memberId, Profile profileRequest);
-
-    void checkNicknameDuplicated(Nickname nicknameRequest);
 }
