+++ src/main/java/com/example/univeus/domain/auth/service/AuthServiceImpl.java
@@ -1,20 +1,13 @@
 package com.example.univeus.domain.auth.service;
 
 
-import static com.example.univeus.common.response.ResponseMessage.*;
-
 import com.example.univeus.domain.auth.GoogleServerLogin.Response;
 import com.example.univeus.domain.auth.SocialLogin;
 import com.example.univeus.domain.auth.TokenProvider;
 import com.example.univeus.domain.auth.dto.UserTokens;
 import com.example.univeus.domain.auth.model.RefreshToken;
-import com.example.univeus.domain.member.exception.MemberException;
-import com.example.univeus.domain.member.model.Department;
-import com.example.univeus.domain.member.model.Gender;
 import com.example.univeus.domain.member.model.Member;
 import com.example.univeus.domain.member.service.MemberService;
-import com.example.univeus.presentation.auth.dto.request.AuthRequest.Nickname;
-import com.example.univeus.presentation.auth.dto.request.AuthRequest.Profile;
 import com.example.univeus.presentation.auth.dto.response.AuthResponse.ResponseTokens;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseCookie;
@@ -39,6 +32,7 @@ public class AuthServiceImpl implements AuthService {
     }
 
 
+    @Override
     public ResponseTokens issueTokens(Long memberId) {
         UserTokens userTokens = tokenProvider.generateTokens(memberId.toString());
         Member member = memberService.findById(memberId);
@@ -58,22 +52,5 @@ public class AuthServiceImpl implements AuthService {
         member.checkProceed();
 
         return issueTokens(member.getId());
-    }
-
-    @Override
-    public void registerProfile(Long memberId, Profile profileRequest) {
-        Department department = Department.of(profileRequest.department());
-        Gender gender = Gender.of(profileRequest.gender());
-
-        memberService.updateProfile(memberId, department, gender, profileRequest.nickname(),
-                profileRequest.studentId());
-    }
-
-    @Override
-    public void checkNicknameDuplicated(Nickname nicknameRequest) {
-        String nickname = nicknameRequest.nickname();
-        if (memberService.findByNickname(nickname).isPresent()) {
-            throw new MemberException(MEMBER_NICKNAME_DUPLICATED);
-        }
     }
 }
