+++ src/main/java/com/example/univeus/domain/member/service/MemberService.java
@@ -3,6 +3,7 @@ package com.example.univeus.domain.member.service;
 import com.example.univeus.domain.member.model.Department;
 import com.example.univeus.domain.member.model.Gender;
 import com.example.univeus.domain.member.model.Member;
+import com.example.univeus.presentation.member.dto.request.MemberRequest;
 import java.util.Optional;
 
 public interface MemberService {
@@ -19,4 +20,8 @@ public interface MemberService {
     Optional<Member> findByNickname(String nickname);
 
     void updatePhoneNumber(Long memberId, String phoneNumber);
+
+    void registerProfile(Long memberId, MemberRequest.Profile profileRequest);
+
+    void checkNicknameDuplicated(MemberRequest.Nickname nicknameRequest);
 }
