+++ src/main/java/com/example/univeus/presentation/member/controller/MemberController.java
@@ -1,5 +1,7 @@
 package com.example.univeus.presentation.member.controller;
 
+import static com.example.univeus.common.response.ResponseMessage.CHECK_NICKNAME_DUPLICATED_SUCCESS;
+import static com.example.univeus.common.response.ResponseMessage.PROFILE_REGISTER_SUCCESS;
 import static com.example.univeus.common.response.ResponseMessage.UPDATE_PHONE_NUMBER_SUCCESS;
 import static com.example.univeus.presentation.member.dto.request.MemberRequest.*;
 
@@ -8,10 +10,13 @@ import com.example.univeus.common.annotation.MemberOnly;
 import com.example.univeus.common.response.Response;
 import com.example.univeus.domain.auth.model.Accessor;
 import com.example.univeus.domain.member.service.MemberService;
+import com.example.univeus.presentation.auth.dto.request.AuthRequest;
+import com.example.univeus.presentation.member.dto.request.MemberRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -27,7 +32,7 @@ public class MemberController {
     @PatchMapping("/phone")
     public ResponseEntity<Response<String>> updatePhoneNumber(
             @Auth Accessor accessor,
-            @RequestBody @Valid PhoneNumber phoneNumberRequest
+            @Valid @RequestBody PhoneNumber phoneNumberRequest
     ) {
         Long memberId = accessor.getMemberId();
         memberService.updatePhoneNumber(memberId, phoneNumberRequest.phoneNumber());
@@ -37,6 +42,34 @@ public class MemberController {
                 .body(Response.success(
                         UPDATE_PHONE_NUMBER_SUCCESS.getCode(),
                         UPDATE_PHONE_NUMBER_SUCCESS.getMessage()
+                ));
+    }
+
+    @MemberOnly
+    @PostMapping("/profile")
+    public ResponseEntity<Response<String>> registerProfile(
+            @Auth Accessor accessor,
+            @Valid @RequestBody MemberRequest.Profile profileRequest
+    ) {
+        memberService.registerProfile(accessor.getMemberId(), profileRequest);
+        return ResponseEntity
+                .ok()
+                .body(Response.success(
+                        PROFILE_REGISTER_SUCCESS.getCode(),
+                        PROFILE_REGISTER_SUCCESS.getMessage()
+                ));
+    }
+
+    @PostMapping("/nickname/duplicated")
+    public ResponseEntity<Response<String>> checkNicknameDuplicated(
+            @Valid @RequestBody MemberRequest.Nickname nicknameRequest
+    ) {
+        memberService.checkNicknameDuplicated(nicknameRequest);
+        return ResponseEntity
+                .ok()
+                .body(Response.success(
+                        CHECK_NICKNAME_DUPLICATED_SUCCESS.getCode(),
+                        CHECK_NICKNAME_DUPLICATED_SUCCESS.getMessage()
                 ));
     }
 }
