+++ src/test/java/com/example/univeus/presentation/member/controller/MemberControllerTest.java
@@ -1,5 +1,7 @@
 package com.example.univeus.presentation.member.controller;
 
+import static com.example.univeus.common.response.ResponseMessage.CHECK_NICKNAME_DUPLICATED_SUCCESS;
+import static com.example.univeus.common.response.ResponseMessage.PROFILE_REGISTER_SUCCESS;
 import static com.example.univeus.common.response.ResponseMessage.UPDATE_PHONE_NUMBER_SUCCESS;
 import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.is;
@@ -7,6 +9,7 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 
@@ -16,12 +19,16 @@ import com.example.univeus.domain.member.service.MemberService;
 import com.example.univeus.presentation.BaseControllerTest;
 import com.example.univeus.presentation.member.dto.request.MemberRequest;
 import com.example.univeus.presentation.member.dto.request.MemberRequest.PhoneNumber;
+import com.example.univeus.presentation.member.dto.request.MemberRequest.Profile;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -103,6 +110,112 @@ class MemberControllerTest extends BaseControllerTest {
                                     is("휴대폰 번호는 공백이 될 수 없습니다."),
                                     is("휴대폰 번호의 길이가 올바르지 않습니다."),
                                     is("휴대폰 번호는 숫자만 포함해야 합니다."))));
+        }
+    }
+
+    @Nested
+    @DisplayName("프로필 등록 테스트")
+    class TestRegisterProfile {
+
+        @Test
+        void 올바른_형식으로_요청을_보낼경우_예외가_발생하지_않는다() throws Exception {
+
+            MemberRequest.Profile profile = Profile.of(
+                    "nickname",
+                    "department",
+                    "gender",
+                    "studentId"
+            );
+
+            ResultActions resultActions =
+                    mockMvc.perform(post("/api/v1/member/profile")
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .content(objectMapper.writeValueAsString(profile)));
+
+            resultActions
+                    .andExpect(status().isOk())
+                    .andExpect(jsonPath(("$.code")).value(PROFILE_REGISTER_SUCCESS.getCode()))
+                    .andExpect(jsonPath(("$.message")).value(PROFILE_REGISTER_SUCCESS.getMessage()));
+        }
+
+        @ParameterizedTest
+        @MethodSource("provideProfiles")
+        void 올바르지_않은_프로필_등록_요청을_보낼경우_예외가_발생한다(
+                String nickname, String department, String gender, String studentId
+        ) throws Exception {
+
+            MemberRequest.Profile profile = Profile.of(
+                    nickname,
+                    department,
+                    gender,
+                    studentId
+            );
+
+            ResultActions resultActions =
+                    mockMvc.perform(post("/api/v1/member/profile")
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .content(objectMapper.writeValueAsString(profile)));
+
+            resultActions
+                    .andExpect(status().isBadRequest())
+                    .andExpect(jsonPath(("$.code")).value("FORMAT-001"));
+        }
+
+        private static Stream<Arguments> provideProfiles() {
+            return Stream.of(
+                    Arguments.of("", "department", "gender", "studentId"),       // 닉네임이 공백
+                    Arguments.of("nickname", "", "gender", "studentId"),        // 학과가 공백
+                    Arguments.of("nickname", "department", "", "studentId"),    // 성별이 공백
+                    Arguments.of("nickname", "department", "gender", "")        // 학번이 공백
+            );
+        }
+
+        @Nested
+        @DisplayName("닉네임 중복 테스트")
+        class TestCheckNicknameDuplicated {
+
+            @Test
+            void 올바른_형식으로_요청을_보낼경우_예외가_발생하지_않는다() throws Exception {
+
+                // given
+                MemberRequest.Nickname nicknameRequest = MemberRequest.Nickname.of("testNickname");
+
+                // when
+                ResultActions resultActions =
+                        mockMvc
+                                .perform(post("/api/v1/member/nickname/duplicated")
+                                        .contentType(MediaType.APPLICATION_JSON)
+                                        .content(objectMapper.writeValueAsString(nicknameRequest)));
+
+                // then
+                resultActions
+                        .andExpect(status().isOk())
+                        .andExpect(jsonPath("$.code").value(CHECK_NICKNAME_DUPLICATED_SUCCESS.getCode()))
+                        .andExpect(jsonPath("$.message").value(CHECK_NICKNAME_DUPLICATED_SUCCESS.getMessage()));
+            }
+
+
+            @ParameterizedTest
+            @ValueSource(strings = {"", " ", "    "})
+            void 닉네임이_올바른_입력이_아닐경우_예외를_던진다(String nickname) throws Exception {
+
+                // given
+                MemberRequest.Nickname nicknameRequest = MemberRequest.Nickname.of(nickname);
+
+                // when
+                ResultActions resultActions =
+                        mockMvc
+                                .perform(post("/api/v1/member/nickname/duplicated")
+                                        .contentType(MediaType.APPLICATION_JSON)
+                                        .content(objectMapper.writeValueAsString(nicknameRequest)));
+
+                // then
+                resultActions
+                        .andExpect(status().isBadRequest())
+                        .andExpect(jsonPath("$.status").value(false))
+                        .andExpect(jsonPath("$.code").value("FORMAT-001"))
+                        .andExpect(jsonPath("$.message").value("닉네임은 공백이 될 수 없습니다."));
+            }
         }
     }
 }
